[BE] feat: 로그인 api 및 entity 수정

+++ backend/src/main/java/kr/remerge/stylehub/domain/cart/entity/CartItem.java
@@ -27,7 +27,7 @@ public class CartItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cart_item_id")
-    private Long cartItemId;
+    private Integer cartItemId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
